requirements.txt
# Use Python 3.11 or 3.12 for this project (scikit-learn has no wheels for 3.14).
# On Windows without C++ Build Tools, use: pip install -r requirements-base.txt
# Then optionally: pip install -r requirements-ml.txt (needs Build Tools for scikit-learn)
# https://visualstudio.microsoft.com/visual-cpp-build-tools/
asgiref==3.8.1
Django==5.2.1
django-browser-reload==1.18.0
django-tailwind==4.0.1
django-model-utils==4.3.1
openpyxl==3.1.2
xhtml2pdf==0.2.11
pdfkit==1.0.0  # Python wrapper for wkhtmltopdf
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.3
tzdata==2025.2
django-extensions
whitenoise==6.7.0
dj-database-url==2.2.0
python-dotenv==1.0.1
django-redis==5.4.0  # For Redis caching (optional - only needed if you add Redis)
celery==5.3.4  # For background task processing
channels==4.0.0  # For WebSocket support (Django Channels)
channels-redis==4.1.0  # Redis backend for Channels
daphne==4.0.0  # ASGI server for WebSocket support
django-storages==1.14.2  # For DigitalOcean Spaces (S3-compatible storage)
boto3==1.34.0  # AWS SDK for Python (used by django-storages for Spaces)
# GDAL==3.10.*  # Commented out - not needed (django.contrib.gis is disabled in settings.py)
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
holidays>=0.40  # Optional: PH holidays for years not in projeng.working_days (2026 is built-in)
